+++ src/UserInterface/PythonNetworkStreamModule.cpp
@@ -756,7 +756,7 @@ PyObject* netSendItemMovePacket(PyObject* poSelf, PyObject* poArgs)
 	}
 
 	CPythonNetworkStream& rkNetStream=CPythonNetworkStream::Instance();
-	rkNetStream.SendItemMovePacket(Cell, ChangeCell, (BYTE) num);
+	rkNetStream.SendItemMovePacket(Cell, ChangeCell, (WORD) num);
 	return Py_BuildNone();
 }
